Factor main% out of GUI module, add grid.
main% is now it's own module, and will be where the meat of the project
is located. I've added deep-map, which is used to update a grid, which
is currently a nested list. Added an example of the hash-table I might
be using to hold entites in the game, where each monster is hashed by
its posn, and adding monsters conses the monster onto the list of other
monsters hashed to that key.