Add more tests, added draw-grid.
The testing module now includes rudimentary tests for triangles.  Added
random-shape functions for the level module and made a a function that
draws a grid of cells by running each cell through a level and coloring
it based on whether it belongs in a room or not. Not sure if it works
correctly.