Fixed grid-rollover bug.
Previously, the grid was not rendering any images correctly because it
was misaligning all cells. First they were backwards, because I'm using
an accumulator to collect images before appending them, and then it was
skipping the first value of every new row because it uses (rest a-grid),
which drops the first value, when recursing on a new row and vertically
appending previous row. So all cells were "rolling over" by a value
every row, causing shapes to render incorrectly. Frequently they were
split in half because the cells "drifted" to the wrong side and rolled
over to the other side.

There's still a bug with the random number generator, where it is
sometimes passed a negative or fractional value. I think it's in the
random triangle function, that's the only function which does anything
more complicated than calling integers directly with random.