Fix invalid random number generation.
The problem was in the y-intercept generation of triangles, which would
generate a number between 1 and the maximum height of a room. This
number was then used to generate a base for the triangle by generating a
new random number between 1 and (- y-intercept 5), which occasionally
spat out zero or negative integers as a maximum bound.